https://github.com/e-government-ua/i/issues/1071 wf-central: finance(pay+merchant), document, subject
