+++ central-js/server/api/organs/index.controller.js
@@ -3,11 +3,11 @@ var _ = require('lodash');
 var activiti = require('../../components/activiti');
 
 module.exports.getSubjectOrganJoin = function (req, res) {
-  activiti.sendGetRequest(req, res, '/services/getSubjectOrganJoins', _.extend(req.query, req.params)); // {nID_SubjectOrgan:1} for test
+  activiti.sendGetRequest(req, res, '/subject/getSubjectOrganJoins', _.extend(req.query, req.params)); // {nID_SubjectOrgan:1} for test
 };
 
 module.exports.getOrganAttributes = function(req, res) {
-  var apiReq = activiti.buildRequest(req, '/services/getSubjectOrganJoinAttributes', _.extend(req.query, req.params));
+  var apiReq = activiti.buildRequest(req, '/subject/getSubjectOrganJoinAttributes', _.extend(req.query, req.params));
   apiReq.body = req.body;
   apiReq.json = true;
   activiti.executePostRequest(apiReq, res);
