+++ central-js/server/api/subject-organ-join-tax/index.controller.js
@@ -2,5 +2,5 @@
 var activiti = require('../../components/activiti');
 
 module.exports.getSubjectOrganJoinTaxList = function (req, res) {
-  activiti.sendGetRequest(req, res, '/services/getSubjectOrganJoinTax', req.query);
+  activiti.sendGetRequest(req, res, '/subject/getSubjectOrganJoinTax', req.query);
 };
