+++ wf-base/src/main/java/org/igov/io/liqpay/LiqBuy.java
@@ -58,7 +58,7 @@ public class LiqBuy {
             oLanguage = DEFAULT_LANG;
         }
 
-        String URI = "/wf/service/merchant/getMerchant";
+        String URI = "/wf/service/finance/getMerchant";
         Map<String, String> paramMerchant = new HashMap<String, String>();
         paramMerchant.put("sID", sID_Merchant);
